[A14] ! Give up on Web frontend.
There are security implications of an artilect hosting their own frontend,
the implications I'm not comfortable accepting.

- Frontend for a specific artilect may defy user expectations. E.g., if
  end-to-end encrypted private chats would be implemented, then such a
  frontend might trivially break the end-to-end encryption guarantees.
  There are many more potential examples of what kind of information
  the cuser would expect to be local-only per the standard artilect
  client, but that a custom web frontend would push onto the backend.
- Implementing artilect proxy service in a way that would support web
  frontend securely is quite difficult. Proxy itself might be MITM in
  that setup.
- The ability to host a custom HTTP client might create incentives to
  deviate from the standard and create incompatible implementations.

...but honestly, there is just a lot complexity and effort to ensure
that web frontend compiles and works the same as desktop/mobile.

Note: This also closes off the embeddable web frontend deployment, which
is unfortunate. So we might go back on this for that feature
specifically, but in that case it'll be built without the expectation
of a general-purpose self-contained web frontend.